v4.1
added Inventory and a function to add stuff into it (still need to add so things inside the inventory can be moved around and make it so the armor slots are actually usable. they are still just a place to put them. they don't change anything in the world like your apperance and the stats on your character.)